01_gcf_get_stock_data_upload_storage/requirements.txt
yfinance
pandas
google-cloud-storage